requirements.txt
wxauto>=3.9.8.15
openai>=1.12.0
httpx>=0.27.0
quart>=0.19.0
hypercorn>=0.16.0
pydantic>=2.0.0
colorlog>=6.8.0
pyyaml>=6.0
schedule>=1.2.0
pyinstaller>=6.0.0
aiosqlite>=0.17.0
tenacity>=8.2.0
chromadb>=0.4.0
Pillow>=11.0.0
coverage>=7.0.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
